requirements.txt
fastapi
pydantic
httpx
python-dotenv
uvicorn
